fix: Remove neighbors if compactTo is unindexed
See also: #36360
